requirements.txt
speechrecognition
pydub
langdetect
fastapi
twilio
python-dotenv
pydantic
requests
